project/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg[binary]>=3.2.10,<3.4
alembic==1.15.2
python-jose[cryptography]==3.4.0
passlib==1.7.4
python-multipart==0.0.20
pydantic-settings==2.8.1
email-validator==2.3.0
numpy
Pillow
reportlab==4.2.5
pytest==8.3.5
httpx==0.28.1
